% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gh_labels}
\alias{gh_labels}
\title{Default GitHub Issue Labels}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{\code{label}}{character. Label names.}
\item{\code{description}}{character. Label descriptions.}
\item{\code{colour}}{character. Label colors represented in hexadecimal format.}
}
}
\usage{
gh_labels
}
\description{
This \code{data.frame} houses my default repository issue labels applied to R projects
via \code{\link[usethis:use_github_labels]{usethis::use_github_labels()}}.
}
\seealso{
\code{\link[=use_gh_labels]{use_gh_labels()}}, \code{\link[usethis:use_github_labels]{usethis::use_github_labels()}}.
}
\keyword{datasets}
