% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_template}
\alias{render_template}
\title{Render Template}
\usage{
render_template(template_path, out_path, data = list(), open = TRUE)
}
\arguments{
\item{template_path}{Path to the template file}

\item{out_path}{Path to desired output location}

\item{data}{list of fields to pass to the template file}

\item{open}{logical - should the file be opened?}
}
\value{
Invisible
}
\description{
This function is a modification of \code{\link[usethis:use_template]{usethis::use_template()}} that applies user
supplied data fields to a given template file via \code{\link[whisker:whisker.render]{whisker::whisker.render()}}.
}
\examples{
\dontrun{
render_template(system.file("templates/pkgdevt-script", package = "templateeR"),
                ".", data = list(package_name = "testPackage"),
                open = TRUE)
}
}
