% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateeR-package.R
\docType{package}
\name{templateeR-package}
\alias{templateeR}
\alias{templateeR-package}
\title{templateeR: Collection of templates for R}
\description{
Apply variouse R related templates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/templateeR}
  \item Report bugs at \url{https://github.com/jimbrig/templateeR/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
