% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_git_cliff.R
\name{test_git_cliff_installated}
\alias{test_git_cliff_installated}
\title{Test that \code{git-cliff} is installed}
\usage{
test_git_cliff_installated()
}
\description{
Test that \code{git-cliff} is installed
}
\keyword{internal}
