% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_git_cliff.R
\name{use_git_cliff}
\alias{use_git_cliff}
\title{Use \code{git-cliff}}
\usage{
use_git_cliff(path = "CHANGELOG.md", config = "cliff.toml")
}
\arguments{
\item{path}{Path to output generated CHANGELOG.md file.}

\item{config}{Path to \code{cliff.toml} configuration file.}
}
\value{
Nothing; used for its side-effects.
}
\description{
\code{git-cliff} is a useful tool for automating the generation of repository
CHANGELOGs and following good practices with git by using the
\href{https://www.conventionalcommits.org}{conventional commits} methodology.

This function implements \code{git-cliff} into your repo by adding a \code{cliff.toml}
default configuration file and then running \code{git-cliff} to output a CHANGELOG.md
file based off the repo's git history.
}
\examples{
\dontrun{
 use_git_cliff()
}

}
\seealso{
\itemize{
\item \href{https://github.com/orhun/git-cliff/}{orhun/git-cliff GitHub repo}
\item \href{https://github.com/orhun/git-cliff/blob/main/config/cliff.toml}{Default cliff.toml config file}
\item \href{https://www.conventionalcommits.org}{Conventional Commits}
}
}
