% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_actions.R
\name{use_git_cliff_action}
\alias{use_git_cliff_action}
\title{Use \code{git-cliff} GitHub Action Workflow}
\usage{
use_git_cliff_action(save_as = "git-cliff.yml", open = FALSE)
}
\arguments{
\item{save_as}{File name to save as; defaults to \code{git-cliff.yml}. Will be saved
in the \code{.github/workflows} directory.}

\item{open}{Logical - should the file be opened?}
}
\value{
Invisible
}
\description{
Adds a github action to auto-generate a \code{CHANGELOG.md} using GitHub actions and
\code{git-cliff}.
}
\examples{
\dontrun{
 use_git_cliff_action()
}
}
\seealso{
\code{\link[=use_git_cliff]{use_git_cliff()}}
}
